+++ Robust.Server/Console/ConGroupController.cs
@@ -1,8 +1,12 @@
 ﻿using Robust.Server.Player;
+using Robust.Shared.Map;
+using Robust.Shared.Players;
+using Robust.Shared.Toolshed;
+using Robust.Shared.Toolshed.Errors;
 
 namespace Robust.Server.Console
 {
-    internal sealed class ConGroupController : IConGroupController
+    internal sealed class ConGroupController : IConGroupController, IPermissionController
     {
         public IConGroupControllerImplementation? Implementation { get; set; }
 
@@ -29,6 +33,12 @@ namespace Robust.Server.Console
         public bool CanAdminReloadPrototypes(IPlayerSession session)
         {
             return Implementation?.CanAdminReloadPrototypes(session) ?? false;
+        }
+
+        public bool CheckInvokable(CommandSpec command, ICommonSession? user, out IConError? error)
+        {
+            error = null;
+            return Implementation?.CheckInvokable(command, user, out error) ?? false;
         }
     }
 }