+++ Robust.Shared/Toolshed/TypeParsers/IAsType.cs
@@ -1,0 +1,9 @@
+﻿namespace Robust.Shared.Toolshed.TypeParsers;
+
+/// <summary>
+/// Generalized unboxing of a value from a containing structure.
+/// </summary>
+public interface IAsType<out T>
+{
+    public T AsType();
+}